Move TwitterStream and Credentials types to their own files and cleanup definitions based on #pullrequestreview-534608395 feedback
