+++ src/TwitterStream.d.ts
@@ -1,0 +1,7 @@
+export class TwitterStream<T> implements AsyncIterable<T> {
+  constructor(connect: () => Promise<any>, close: () => void);
+
+  [Symbol.asyncIterator](): AsyncIterator<T>;
+
+  public close(): void;
+}
